Uploader 1.1 improve ments

+++ components/divami_components/web_worker/fileUploadManager.ts
@@ -14,38 +14,45 @@ self.onmessage = async (event: MessageEvent<eventType>) => {
     if(filesList?.length===0) {
         return;
     }
-    
     let currentTask = 0
     for (currentTask = 0; currentTask < Math.min(5, filesList.length); currentTask++) {
-        uploadTask(currentTask);    
+         uploadTask(currentTask);    
     }
     self.postMessage({filesList,completedFileList})
      
 
     async function uploadTask(i: number) {
-        await axios.put( `${API.BASE_URL}/rawimages/${filesList[i].uploadObject._id}`,{status: 'Started'}, {
-            headers: event.data.headerValue,
-          }).then((response) => {
-            return response.data;
-          })
-          .catch((error) => {
-            return error.response;
-          });
-        const worker = new Worker(new URL('./fileUploaderWorker.ts', import.meta.url));
-        worker.postMessage(filesList[i]);
-        worker.onmessage = (event) => {
-            const { status, id, errorMessage } = event.data;
-            let file = filesList.find((e) => { return e.uploadObject._id === id; });
-            if (file) {
-                file.status = status;
-                file.error = errorMessage;
-                file.progress.value = status === UploadStatus.success ? 100 : 0;
-                completedFileList.push(file);
-                self.postMessage({ filesList, completedFileList });
-
+        if(filesList[i].status != 1)
+        { 
+            if(filesList[i].status === 2)
+            {
+                filesList[i].status = 0
             }
-            if(currentTask < filesList.length) uploadTask(currentTask++)
-            worker.terminate();
-        };
+            await axios.put( `${API.BASE_URL}/rawimages/${filesList[i].uploadObject._id}`,{status: 'Started'}, {
+                headers: event.data.headerValue,
+              }).then((response) => {
+                return response.data;
+              })
+              .catch((error) => {
+                return error.response;
+              });
+            const worker = new Worker(new URL('./fileUploaderWorker.ts', import.meta.url));
+            worker.postMessage(filesList[i]);
+            worker.onmessage = (event) => {
+                const { status, id, errorMessage } = event.data;
+                let file = filesList.find((e) => { return e.uploadObject._id === id; });
+                if (file) {
+                    file.status = status;
+                    file.error = errorMessage;
+                    file.progress.value = status === UploadStatus.success ? 100 : 0;
+                    completedFileList.push(file);
+                    self.postMessage({ filesList, completedFileList });
+    
+                }
+                if(currentTask < filesList.length) uploadTask(currentTask++)
+                worker.terminate();
+            };
+        }
+        
     }
 }
